+++ src/renderable/container.js
@@ -227,9 +227,22 @@ let globalFloatingCounter = 0;
             this.updateBounds();
         }
 
-        // if a physic body is defined, add it to the game world
-        if (child.body instanceof Body) {
-            this.getRootAncestor().addBody(child.body);
+        // if a physic body(ies) to the game world
+        if (this.isAttachedToRoot()) {
+            var worldContainer = this.getRootAncestor();
+            if (child.body instanceof Body) {
+                worldContainer.addBody(child.body);
+            }
+            // if the child is a container
+            if (child instanceof Container) {
+                // add all container child bodies
+                // TODO: make it recursive ?
+                child.forEach((cchild) => {
+                    if (cchild.body instanceof Body) {
+                        worldContainer.addBody(cchild.body);
+                    }
+                });
+            }
         }
 
         // triggered callback if defined
@@ -276,9 +289,22 @@ let globalFloatingCounter = 0;
                 this.updateBounds();
             }
 
-            // if a physic body is defined, add it to the game world
-            if (child.body instanceof Body) {
-                this.getRootAncestor().addBody(child.body);
+            // if a physic body(ies) to the game world
+            if (this.isAttachedToRoot()) {
+                var worldContainer = this.getRootAncestor();
+                if (child.body instanceof Body) {
+                    worldContainer.addBody(child.body);
+                }
+                // if the child is a container
+                if (child instanceof Container) {
+                    // add all container child bodies
+                    // TODO: make it recursive ?
+                    child.forEach((cchild) => {
+                        if (cchild.body instanceof Body) {
+                            worldContainer.addBody(cchild.body);
+                        }
+                    });
+                }
             }
 
             // triggered callback if defined